api/requirements.txt
django<3.3
djangorestframework
django-environ
psycopg2
watchdog[watchmedo]
drf-spectacular
jinja2
whitenoise
django-cors-headers
django-rest-knox
sentry-sdk
